Feature/fix bug chi tiet lich hop

See merge request quannv_edsolabs/ccvc-mobile!452

+++ lib/domain/model/lich_hop/chi_tiet_lich_hop_model.dart
@@ -289,7 +289,7 @@ class file {
   final bool? isPrivate;
   final String? name;
   final String? path;
-  final int? size;
+  final double? size;
   final String? updatedAt;
   final String? updatedBy;
